+++ src/axes.ts
@@ -607,14 +607,21 @@ export class Axis extends Shape {
       maxValue + valueRange * this.marginRatio];
   }
 
+  public setRubberbandRange(minValue: number, maxValue: number): void {
+    this.rubberBand.minValue = Math.min(minValue, maxValue);
+    this.rubberBand.maxValue = Math.max(minValue, maxValue);
+  }
+
+  public setRubberBandRangeDebug(minValue: string, maxValue: string): void {
+    this.setRubberbandRange(Number(minValue), Number(maxValue));
+  }
+
   private updateRubberBand(): void {
     const canvasCoords = new Vertex(
       this.relativeToAbsolute(this.rubberBand.minValue),
       this.relativeToAbsolute(this.rubberBand.maxValue)
     );
     this.rubberBand.updateCoords(canvasCoords, this.origin, this.end);
-    [this.rubberBand.origin, this.rubberBand.size] = this.rubberBand.computeRectProperties(this.origin);
-    this.rubberBand.buildPath();
   }
 
   public drawRubberBand(context: CanvasRenderingContext2D, canvasMatrix: DOMMatrix): void {
@@ -655,11 +662,11 @@ export class Axis extends Shape {
     if (this.isClicked) {
       if (this.title.isClicked) this.mouseMoveClickedTitle(mouseCoords)
       else this.mouseMoveClickedArrow(mouseCoords);
+      this.rubberBand.updateMinMaxValueOnMouseMove(
+        this.absoluteToRelative(this.rubberBand.canvasMin),
+        this.absoluteToRelative(this.rubberBand.canvasMax)
+      );
     }
-    this.rubberBand.updateMinMaxValueOnMouseMove(
-      this.absoluteToRelative(this.rubberBand.canvasMin),
-      this.absoluteToRelative(this.rubberBand.canvasMax)
-    );
   }
 
   public mouseDown(mouseDown: Vertex): void {