+++ src/interactions.ts
@@ -1,4 +1,5 @@
 import { Subject } from "rxjs"
+import { RubberBand } from "./shapes";
 
 export interface HighlightData {
   referencePath: string,
@@ -6,4 +7,10 @@ export interface HighlightData {
   select: boolean
 }
 
+export interface FilterUpdate {
+  id: string,
+  rubberbands: Map<string, RubberBand>
+}
+
 export const highlightShape: Subject<HighlightData> = new Subject();
+export const filterUpdate: Subject<FilterUpdate> = new Subject();