+++ src/remoteFigure.ts
@@ -9,6 +9,7 @@ import { Axis } from "./axes"
 import { PointSet, ShapeCollection, GroupCollection } from "./collections"
 import { DataInterface } from "./dataInterfaces"
 
+
 export class RemoteFigure extends Rect {
   public context: CanvasRenderingContext2D;
   public axes: Axis[] = [];
@@ -187,6 +188,21 @@ export class RemoteFigure extends Rect {
       if (axis.rubberBand.length != 0) axesSelections.push(this.updateSelected(axis));
     })
     this.updateSelection(axesSelections);
+  }
+
+  public setFeatureFilter(feature: string, minValue: number, maxValue: number): boolean {
+    for (const axe of this.axes) { 
+      if (axe.name.toLowerCase() == feature.toLowerCase()) {
+        axe.setRubberbandRange(minValue, maxValue);
+        this.draw();
+        return true
+      }
+    }
+    return false
+  }
+
+  public setFeatureFilterDebug(feature: string, minValue: number, maxValue: number): boolean {
+    return this.setFeatureFilter(feature, Number(minValue), Number(maxValue))
   }
 
   public changeAxisFeature(name: string, index: number): void {